backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
openai-agents
google-genai
requests
python-dotenv
pydantic